IK Implemented (still not quite working...)
A few things:

1. I've slowed down the time delay in graphing to better see the elbow
up/elbow down poses and compare them. It seems for theta1_1, we're
close, but for theta1_2, we're far.
2. I changed the theta1 equations, based on the book's version for
offset, so potentially a typo here?
3. Pose order is four combinations of arm positions with one set of
wrist orientations, then the same four arm positions with the second set
of wrist orientations.
4. We were backing out from the tip to the wrist center AFTER
calculating the arm's position, when we in fact needed to use the wrist
center as the "tip" for kinematic decoupling purposes, so I changed
that.
5. Rotation matrices and euler angles are now automatically calculated
for each arm pose! (main exciting point... save the best for last!!)