Wrist singularity and partial FK/IK/FK loop tests
Got the special cases in for the wrist singularities.  Partially
implemented a new test case to generate random joint angles, run FK, run
IK using the position/orientation from FK, then rerun FK and compare to
the original results.  Something weird is happening in the rotation
portion, if you inspect the before/after Rs they are the same with a few
signs swapped.  It's probably something obvious and I've been staring at
it too long, but it's not crucial since it's not required anyway.  I'll
take another look tomorrow afternoon unless one of you beats me to it.